Refs #29597 - refactor task actions folders
