+++ webpack/ForemanTasks/Components/TaskActions/TaskActionsConstants.js
@@ -1,0 +1,11 @@
+export const FOREMAN_TASK_DETAILS_TOGGLE_UNLOCK_MODAL =
+  'FOREMAN_TASK_DETAILS_TOGGLE_UNLOCK_MODAL';
+export const FOREMAN_TASK_DETAILS_TOGGLE_FORCE_UNLOCK_MODAL =
+  'FOREMAN_TASK_DETAILS_TOGGLE_FORCE_UNLOCK_MODAL';
+
+export const TASKS_RESUME_REQUEST = 'TASKS_RESUME_REQUEST';
+export const TASKS_RESUME_SUCCESS = 'TASKS_RESUME_SUCCESS';
+export const TASKS_RESUME_FAILURE = 'TASKS_RESUME_FAILURE';
+export const TASKS_CANCEL_REQUEST = 'TASKS_CANCEL_REQUEST';
+export const TASKS_CANCEL_SUCCESS = 'TASKS_CANCEL_SUCCESS';
+export const TASKS_CANCEL_FAILURE = 'TASKS_CANCEL_FAILURE';
