+++ webpack/ForemanTasks/Components/TasksTable/TasksTableActions.js
@@ -1,8 +1,6 @@
 import { getURIQuery } from 'foremanReact/common/helpers';
 import { getTableItemsAction } from 'foremanReact/components/common/table';
-import API from 'foremanReact/API';
-import { addToast } from 'foremanReact/redux/actions/toasts';
-import { sprintf } from 'foremanReact/common/I18n';
+
 import {
   TASKS_TABLE_ID,
   SELECT_ROWS,
@@ -11,21 +9,10 @@ import {
   UNSELECT_ROWS,
   UPDATE_CLICKED,
   OPEN_SELECT_ALL,
-  TASKS_RESUME_REQUEST,
-  TASKS_RESUME_SUCCESS,
-  TASKS_RESUME_FAILURE,
-  TASKS_CANCEL_REQUEST,
-  TASKS_CANCEL_SUCCESS,
-  TASKS_CANCEL_FAILURE,
 } from './TasksTableConstants';
-import { TOAST_TYPES } from '../common/ToastTypesConstants';
 import { getApiPathname } from './TasksTableHelpers';
 import { fetchTasksSummary } from '../TasksDashboard/TasksDashboardActions';
-import {
-  resumeToastInfo,
-  cancelToastInfo,
-  toastDispatch,
-} from './TasksTableActionHelpers';
+import { cancelTaskRequest, resumeTaskRequest } from '../TaskActions';
 
 export const getTableItems = url =>
   getTableItemsAction(TASKS_TABLE_ID, getURIQuery(url), getApiPathname(url));
@@ -45,34 +32,6 @@ export const cancelTask = ({
   reloadPage(url, parentTaskID, dispatch);
 };
 
-export const cancelTaskRequest = (id, name) => async dispatch => {
-  dispatch(
-    addToast({
-      type: TOAST_TYPES.INFO,
-      message: sprintf('Trying to cancel %s task', name),
-    })
-  );
-  dispatch({ type: TASKS_CANCEL_REQUEST });
-  try {
-    await API.post(`/foreman_tasks/tasks/${id}/cancel`);
-    dispatch({ type: TASKS_CANCEL_SUCCESS });
-    toastDispatch({
-      type: 'cancelled',
-      name,
-      toastInfo: cancelToastInfo,
-      dispatch,
-    });
-  } catch ({ response }) {
-    dispatch({ type: TASKS_CANCEL_FAILURE, payload: response });
-    toastDispatch({
-      type: 'skipped',
-      name,
-      toastInfo: cancelToastInfo,
-      dispatch,
-    });
-  }
-};
-
 export const resumeTask = ({
   taskId,
   taskName,
@@ -81,29 +40,6 @@ export const resumeTask = ({
 }) => async dispatch => {
   await dispatch(resumeTaskRequest(taskId, taskName));
   reloadPage(url, parentTaskID, dispatch);
-};
-
-export const resumeTaskRequest = (id, name) => async dispatch => {
-  dispatch({ type: TASKS_RESUME_REQUEST });
-  try {
-    await API.post(`/foreman_tasks/tasks/${id}/resume`);
-
-    dispatch({ type: TASKS_RESUME_SUCCESS });
-    toastDispatch({
-      type: 'resumed',
-      name,
-      toastInfo: resumeToastInfo,
-      dispatch,
-    });
-  } catch ({ response }) {
-    dispatch({ type: TASKS_RESUME_FAILURE, payload: response });
-    toastDispatch({
-      type: 'failed',
-      name,
-      toastInfo: resumeToastInfo,
-      dispatch,
-    });
-  }
 };
 
 export const selectPage = results => dispatch => {
