+++ webpack/ForemanTasks/Components/TasksTable/TasksTableConstants.js
@@ -8,13 +8,6 @@ export const UNSELECT_ALL_ROWS = 'UNSELECT_ALL_ROWS';
 export const SELECT_ALL_ROWS = 'SELECT_ALL_ROWS';
 export const OPEN_SELECT_ALL = 'OPEN_SELECT_ALL';
 
-export const TASKS_RESUME_REQUEST = 'TASKS_RESUME_REQUEST';
-export const TASKS_RESUME_SUCCESS = 'TASKS_RESUME_SUCCESS';
-export const TASKS_RESUME_FAILURE = 'TASKS_RESUME_FAILURE';
-export const TASKS_CANCEL_REQUEST = 'TASKS_CANCEL_REQUEST';
-export const TASKS_CANCEL_SUCCESS = 'TASKS_CANCEL_SUCCESS';
-export const TASKS_CANCEL_FAILURE = 'TASKS_CANCEL_FAILURE';
-
 export const BULK_CANCEL_PATH = 'bulk_cancel';
 export const BULK_RESUME_PATH = 'bulk_resume';
